#644 Split mapstatetoprops
